fix: Add dao implementation for profile

+++ lab2/src/main/java/com/example/lab2/dao/AbstractDao.java
@@ -3,7 +3,7 @@ package com.example.lab2.dao;
 import java.util.Collection;
 
 public interface AbstractDao<T> {
-    T get(Integer id);
+    T get(Long id);
 
     Collection<T> findAll();
